+++ spec/unit/recipes/default_spec.rb
@@ -7,11 +7,11 @@
 require 'spec_helper'
 
 describe 'managed_directory::default' do
-  context 'When all attributes are default, on CentOS 6.6' do
+  context 'When all attributes are default, on CentOS 6.8' do
     let(:chef_run) do
       runner = ChefSpec::SoloRunner.new(
         platform: 'centos',
-        version: '6.6'
+        version: '6.8'
       ) do |node|
         # Set/override attributes here
         # node.automatic['virtualization']['role'] = 'guest'